+++ src/app.ts
@@ -1,27 +1,6 @@
 import fastify from 'fastify'
-import { z } from 'zod'
-import { prisma } from './lib/prisma'
+import { appRoutes } from './http/routes'
 
 export const app = fastify()
 
-app.post('/users', async (req, res) => {
-  const registerBodySchema = z.object({
-    name: z.string(),
-    email: z.string().email(),
-    password: z.string().min(6),
-    avatar: z.string().optional().default(''),
-  })
-
-  const { name, email, password, avatar } = registerBodySchema.parse(req.body)
-
-  await prisma.user.create({
-    data: {
-      name,
-      email,
-      password_hash: password,
-      avatar,
-    },
-  })
-
-  return res.status(201).send()
-})
+app.register(appRoutes)
